% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_codons.R
\name{give_codons}
\alias{give_codons}
\title{give_codons}
\usage{
give_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{the DNA or RNA sequence that we are working with}

\item{start}{a starting position in the sequence that indicates where to start reading}
}
\value{
returns all the codons
}
\description{
give_codons
}
\examples{
give_codons("ATGCCCGTA")
}
