% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_seq.R
\name{random_seq}
\alias{random_seq}
\title{random_seq}
\usage{
random_seq(base_pairs)
}
\arguments{
\item{base_pairs}{Number of base pairs for the sequence to have.}
}
\value{
The functions return a random DNA sequence of length determined by the user's input (base_pairs)
}
\description{
random_seq
}
\examples{
random_seq(9)
}
