% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe}
\alias{transcribe}
\title{TRANSCRIBE}
\usage{
transcribe(DNA)
}
\arguments{
\item{DNA}{It uses DNA as input}
}
\value{
It returns the ARN sequence corresponding to the inputed DNA
}
\description{
TRANSCRIBE
}
