% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_density.R
\name{base_density}
\alias{base_density}
\title{base_density plot}
\usage{
base_density(sequence)
}
\arguments{
\item{sequence}{DNA or RNA Sequence to plot}
}
\value{
The function plots the count of bases present in your sequence
}
\description{
base_density plot
}
